Adding function for user to control rotation of spaceship
Did a lot of testing and debugging because found bug where game loop would get laggier and laggier as time went on because of not resetting and close paths in each game object. Also added user input like clicks on left and right hand side of screen which control the rotation of the spaceship.