Refactoring and added logic
Added a fire method to the spaceship game object class to spawn particles at the nose of the triangle. I think my algorithm is correct to calculate the angle to shoot out the particles. Now just have to determine x and y coords of the new nose of the spaceship when it is rotated off center.