Mostly refactoring
Added ScreenDrawer class which holds the responsibility of drawing all the game objects instead of having the GameHolder do it. Changed definition of GameObject interface so now a pass in screenWidth and screenHeight is required for the update function. Made new utility function to return back a region based on a passed in path. Use that function to detmine click regions for left and right circles. ParticleEmitter is now singleton, seemed obvious.