Tested sample Wave using Observable design pattern between Wave and MeteorShower
Created a sample wave in GameHolder. Updated MeteorShower to take in parameters to make it random and also made it extend Observable. Updated ScreenDrawer to also include a unRegisterGameObject method now. Wave class now implements Observer to observe MeteorShower and when MeteorShower is done. Also at the beginning of Wave, text appears showing the Wave name for 3 seconds.