Split up Particle into two subclasses. Improved on Observer and Observable classes in app.
ParticleEmitter now extends Observable and GameHolder observes it so ParticleEmitter can tell GameHolder how many meteors are on screen at all times to signal the end of a wave. And to also tell GameHolder when a meteor was hit to allow GameHolder to increase user score. Particle is split into two subclasses: Laser and Meteor. This is because they both have different sizes, and because they both have different offscreen() methods. Other small refactoring here and there.