Lots of changes
GameActivity observes GameController so GameController can tell GameActivity when to switch back to the main menu in the case of game over and game done. GameController is now singleton and game constants have been added in as well as initializing of all the waves, and new function to start a new wave. Updated observer update() function to determine the game status at all times. onTouchEvent now back in GamePanel instead of GameController, seemed better. ParticleEmitter has function to remove all of its particles. checkCollisions and update functions now refactored and updated to allow for the removing of particles whether they are because of a collision or offscreen. Increased rotation speed of spaceship. Insured Utility class was static, so made constructor private. Wave now notifies observer with enum instead of string.