Changed look of earth (probably will change in near future) and completely changed meteors
Meteors before were simple rectangles painted on canvas, now they are bitmaps. Particle class had to be refactored to not extend rect anymore. So now laser and meteor have draw functions of their own. Made meteor intersect class which takes in a laser parameter and tests whether they intersect using their rects. Wave, now when generating random meteor, also generates random # between 1 and 3 to determine which bitmap to use for the meteor, and it generates a random scale between 0.5 and 0.8 to scale the meteor bitmap.